requirements.txt
asgiref==3.3.1
Django==3.1.4
django-import-export==2.4.0
dj-database-url==0.5.0
pytz==2020.5
PyYAML==5.3.1
sqlparse==0.4.1
psycopg2==2.8.6
gunicorn==20.0.4
